feat(#408): Jahresurlaubsanspruch mit Nachkommastelle (User.vacation_days dezimal) (#411)
Kundenreport (philvdb): eine 3-Tage-Teilzeitkraft hat 28×3/5 = 16,8 Urlaubstage.
Bisher war der Jahresanspruch nur ganzzahlig eingebbar → Rundung auf 17 → 0,2
Tage Ungerechtigkeit vs. einer Kollegin mit unterjährigem Eintritt.

- Migration 066: users.vacation_days Integer → Numeric(4,1) (up→down→up auf PG18
  verifiziert; Spalte = numeric(4,1); 16,8 speichert als 16,8).
- Schema: vacation_days int → float (UserBase/UserUpdate/UserListResponse,
  ge=0/le=50); suggested_vacation_days bleibt int (30×wd/5, immer ganzzahlig).
- Frontend UserForm: Eingabe parseInt → parseFloat + step="0.1".
- Der Calc rechnete bereits mit Decimal(str(user.vacation_days)) — unverändert.

DSGVO-Export-Fix (Folge der Numeric-Umstellung, Review-CRITICAL): zwei rohe
json.dumps-Export-Pfade schrieben vacation_days ungecastet als Decimal →
"Object of type Decimal is not JSON serializable" (HTTP 500 für JEDEN Nutzer,
auch bei ganzzahligem Wert):
- lifecycle_service._user_dict (Art.15-Selbstexport + Admin-Tenant-Export)
- auth.py /me/export (Art.20-Datenübertragbarkeit)
Beide casten jetzt float(...) — exakt wie das benachbarte weekly_hours (auch
Numeric). Vorher war vacation_days Integer, daher kein Cast nötig.

Verifiziert: Backend-Voll-Suite 1378 passed (die 6 verbleibenden Fails sind
pre-existing shift_planning-„MyToday"-Tests, date-abhängig, failen auch auf
master); Frontend tsc + UserForm 13/13 + build; PG18 Migration up→down→up +
numeric(4,1)-Round-Trip. Neuer Test tests/test_vacation_days_decimal.py.
Multi-Agent-Review: 1 CRITICAL (der Export-Decimal-Crash) gefunden + gefixt.

+++ backend/app/models/user.py
@@ -26,7 +26,7 @@ class User(Base):
     last_name = Column(String(100), nullable=False)
     role = Column(Enum(UserRole), default=UserRole.EMPLOYEE, nullable=False)
     weekly_hours = Column(Numeric(4, 1), nullable=False)  # e.g., 20.0, 30.0, 38.5
-    vacation_days = Column(Integer, nullable=False, default=30)
+    vacation_days = Column(Numeric(4, 1), nullable=False, default=30)  # #408: Dezimal (z. B. 16,8)
     work_days_per_week = Column(Integer, nullable=False, default=5)
     track_hours = Column(Boolean, default=True, nullable=False)  # Track Soll/Ist hours for this user
     calendar_color = Column(String(7), nullable=False, default='#93C5FD')  # Pastel blue default
